deduplicate tournament/swiss search into shared pagination module
Move searchButton and searchInput from separate search.ts files in
tournament and swiss into the shared lib/view/pagination module,
eliminating duplicated code.

Entire-Checkpoint: ff7d63c03877
