+++ ui/lib/src/view/pagination.ts
@@ -1,6 +1,7 @@
 import { h, type VNode } from 'snabbdom';
 import * as licon from '../licon';
-import { bind, type MaybeVNodes } from './snabbdom';
+import { bind, onInsert, type MaybeVNodes } from './snabbdom';
+import { userComplete, type UserCompleteOpts } from './userComplete';
 
 export const maxPerPage = 10;
 
@@ -16,9 +17,12 @@ export interface PaginatedCtrl<A> {
   page: number;
   searching: boolean;
   redraw: () => void;
-  data: { nbPlayers: number; me?: { rank: number } };
+  data: { id: string; nbPlayers: number; me?: { rank: number } };
   pages: Record<number, A[]>;
   toggleFocusOnMe(): void;
+  toggleSearch(): void;
+  jumpToPageOf(name: string): void;
+  jumpToRank(rank: number): void;
   userSetPage(page: number): void;
   userPrevPage(): void;
   userNextPage(): void;
@@ -101,4 +105,43 @@ export function pagerData<A>(ctrl: PaginatedCtrl<A>): PagerData<A> {
 
 export function myPage(ctrl: PaginatedCtrl<unknown>): number | undefined {
   return ctrl.data.me ? Math.floor((ctrl.data.me.rank - 1) / 10) + 1 : undefined;
+}
+
+export function searchButton(ctrl: PaginatedCtrl<unknown>): VNode {
+  return h('button.fbt', {
+    class: { active: ctrl.searching },
+    attrs: { 'data-icon': ctrl.searching ? licon.X : licon.Search, title: 'Search tournament players' },
+    hook: bind('click', ctrl.toggleSearch, ctrl.redraw),
+  });
+}
+
+export function searchInput(ctrl: PaginatedCtrl<unknown>, completeOpts: Partial<UserCompleteOpts>): VNode {
+  return h(
+    'div.search',
+    h('input', {
+      attrs: { spellcheck: 'false' },
+      hook: onInsert((el: HTMLInputElement) => {
+        userComplete({
+          input: el,
+          tag: 'span',
+          focus: true,
+          onSelect(v) {
+            ctrl.jumpToPageOf(v.id);
+            ctrl.redraw();
+          },
+          ...completeOpts,
+        });
+        $(el).on('keydown', e => {
+          if (e.code === 'Enter') {
+            const rank = parseInt(e.target.value.replace('#', '').trim());
+            if (rank > 0) ctrl.jumpToRank(rank);
+          }
+          if (e.code === 'Escape') {
+            ctrl.toggleSearch();
+            ctrl.redraw();
+          }
+        });
+      }),
+    }),
+  );
 }
